Implement untetherAll as calls to stopTethering
(cherry-pick of 7266878f8ebb5fb2c281b660f6f04c86ed6d5357 [formerly 5df6ed4a878c778971f388e3cb88e53b8db43b35])

Calling untether() on an interface name leaves the system
in an inconsistent state, since this is an API for use by
network subsystems to start or stop the tethering of a prepared
downstream network interface.  The proper way to disable tethering
is to call stopTethering() with an appropriate technology.

untetherAll() is called in parts of the system UI to disable
all tethering on system state changes (e.g. on the enabling
of data saver).

Bug: 31405407
Test: Enabling data saver now correctly disables tethering
Change-Id: I8e5724a31e015721910a8d278e100daf697c34d7

Former-commit-id: fa6a48661576d2bd93b5e817d3629cca0a50e6c6