Fix Hotspot tile is shown on Guest user
Last tiles are cached in QSTileHost and they are re-used
for other users. But when the cached tile is added, QSTileHost
doesn't check whether it is available or not. So Hotspot tile
is shown on Guest user. To fix this issue, QSTileHost should
check whether the tile is available when it is added.

Bug: 32817349
Test: manual - add Hotspot tile to default tiles
               and switch to Guest user

Change-Id: I96d5876024f6d519bf72a462b5c79dcf41b253f9

Former-commit-id: efd1c54a400de12b5a73983418ce23ee6a47ab6b