Fix code for new malloc debug enable.
When using am dumpheap -n, modify the code to print the new
way malloc debug works instead of the old way.

Also, the old code had a hard-coded number of backtrace frames. The
new code allows this to be variable so the code needed to be modified
for this too.

Test: am dumpheap -n <PID> <FILE> dumps properly.

Change-Id: Id79e4acbe04313090eb2b973f023e5b27591bb7d

Former-commit-id: e336eea49ad3c4480dfd33f0b6332e4c2683bbe0