Reset mCurrentUserId when fingerprintd died
A template fingerprint data is not loaded in fingerprintd when
fingerprintd is died and restarted.
FigerprintService tries to set userId after a restart of fingerprintd,
but it's skipped because the service tries to set same userId already
set to mCurrentUserId.

To load correct template fingerprint data, reset mCurrentUserId
when fingerprintd died.

Bug: 31734514
Change-Id: I6b8d4b5b685765fe33ef8e654593b615240748a8

Former-commit-id: be11bc5f0a95e00a63018aa758d86eb253bab3b2