Eliminate unnecessary object churn in job scheduler updating
The arbitrary-caller "what job is currently running?" state inspector
returns a clone of the JobStatus record for it, not a live reference
to the scheduler's active state.  That's appropriate for most cases
but when doing its own internal bookkeeping, the object churn in cloning
(especially, an unneeded clone just for purposes of comparing to some
other job!) is merely a huge waste of CPU and GC load.

So now we don't do that unnecessarily, and everything is much leaner.

Bug 31751985

Change-Id: Idacb52e9e5ca17b9e12b7ccd2fc2e1ec8b3547f6

Former-commit-id: eafb535072698b0559d8919db4b18618359c11b9