Correct recently introduced logic error in seamless rotation detection.
The code detecting whether we should rotate seamlessly (which for
example we do not support for 180 degree rotations) was moved to before
mRotation was updated, so it was always passing the same value for old
and new rotation, and we were failing to perform the 180 degree rotation
check. Correct this.

Test: Rotate the device 180 degrees at once while in Camera. Things look ok.
Bug: 32025371
Change-Id: I0ef08e8cc4e5d87fdff6d839381196c2b52e71d5

Former-commit-id: e441e7f15f28c4281e18eba339755344809ce6e7