Search for RRO in Java both in SKU subdir and in vendor/overlay
Currently, if the "ro.boot.vendor.overlay.subdir" property is set, we
only search in the subdir. It seems more powerful to search both in
there and in the standard /vendor/overlay directory.

Also, RRO packages have the "priority" attribute in the <overlay>
attribute in AndroidManifest.xml which serves for disambiguation.

Bug: 32202741
Change-Id: I9ef44b8ca4ac0e92ba02e25b3672c6b274fe4639

Former-commit-id: 4a6e2466713e50d642b375b621265013d73e5d8e