Allow DisplayContent children to be removed on display removal
DisplayContent has direct children that we treat as permanent and as
such we don't allow them to be removed. However, this causes an issue
when removing the display as all children of the display are expected
to be removed at that point.
We now allow the removal of the chilren when the display is being removed
and still disallow it in all other cases.

Bug: 32193212
Test: cts-tradefed run commandAndExit cts-dev -m CtsDisplayTestCases
Change-Id: I9a2eeb2418311a9c5931797ab7a831deead3781e

Former-commit-id: 601a3f085eebc55968805b585906a3e8bad4c3ef