Zygote : Block SIGCHLD during fork.
We close the android logging related sockets prior as late as possible
before every fork to avoid having to whitelist them. If one of the
zygote's children dies after this point (but prior to the fork), we can
end up reopening the logging sockets from the SIGCHLD signal handler.

To prevent this from happening, block SIGCHLD during this critical
section.

Bug: 32693692
Test: Manual

Change-Id: Ia4763a0bfd821a25b68aba3939cb0a6cbf248de0

Former-commit-id: e9a525829a354c92983a35455ccab16d1b0d3892