Add a flag for whether non-system apps can be installed on internal.
This is for a feature in Android Auto Embedded. Cars have 10+ years of
life and the eMMC has limit number of erase/write cycles. To protect the
eMMC internal storage, we recommend car OEMs to adopt
an SD card and install 3rd party apps to the adopted storage.

We still respect the installLocation flag in AndroidManifest. If the
developer request internalOnly and the OEM does not allow 3rd party apps
on internal, the installation will fail.

Test: Unit tests added

Bug: 30980219
Change-Id: I9c40a099b40264a53b1d9d6ac7d5178bef478ed8

Former-commit-id: 199b5d6efdf559db32b52f4d822de1e3ed6724fe