API: Rename Collections.singleton*() type arg E -> T
This rename is part of the update of Collections.java to OpenJDK8u60.

This is compatible change because generic type argument names affect
neither compile time nor runtime compatibility.

The build system (incorrectly) flags this as a change to a released
API. Removing or changing released public APIs is infeasible on AOSP.
Therefore this change will remain restricted to internal branches
until the API is released.

Bug: 31902309
Test: make
Change-Id: I54bc3d6321429ff18c6057332b469286e8b1d42d

Former-commit-id: be7e5900b2228400399a07c2e66188e0356a6d71