IP connectivity metrics: introduce ConnectStatistics and DNS statistics
This patch introduces a ConnectStatistics message for collecting
Android devices connect() statistics.

It also changes DNS latency events to be grouped by (query_type,
query_return_value) to optimize space usage. For this purpose, a new
DNSLookupLatencies message is introduced and DNSLookupBatch is
deprecated. New counter fields are added for the number of queries,
and number of responses with A and AAAA records.

Test: manually taking a diff with reference ipconnectivity.proto
Bug: 32198976
Change-Id: I6cccd539d2c3d703f2d6eb0c56f54cdcc1631a6c

Former-commit-id: 7c3a7869218216fceccb7fba983bc022e556df21