PackageManager: Fix incorrect usage of flags.
assertPackageIsValid was checking policyFlags (which are PackageParser
parseFlags) for flags defined in the SCAN_* values. This was always
subtly and non-visibly broken, but turned into a full scale breakage
when change dde4ea6e0652597b0fdd1f20d082067ab0666dd2 [formerly 5b29ecad4856622422eb903e34499506b2fb3e93] was merged, and
one of the flags that *happened* to have the same value in the PARSE_
and SCAN_ flagset changed.

Test: make, boot, update system app, update playstore app.
Bug: 33124572
Bug: 33126202

Change-Id: I482f420b2a9c8af38b17d174f1a2ec36c762f58e

Former-commit-id: 84e347f01fa3a0675c3a8d6cc973228a4563b20d