Consider priority when launching a USB app
If there are differnet priorities only consider the acitities with the
highest priorities. Esp, if there is only one app with the highest
priority launch it directly.

This enables an interesting use case:
If a priviledged app wants to always be launched for a device it can
just set the priority to >0. As it is also on system/ the user
_will_not_ get a confirmation dialog and the app will always be directly
launched.

Test: Added app with higher priority to device and saw it to get
auto-selected.
Fixes: 28595282

Change-Id: Ia2c9afa00b5a6e8a00b30a01442da62dd0e33961

Former-commit-id: 733d322cf594bb29a57e5e44c5c8100c4cf3ffcc