Don't change visibility of fragment Views that aren't visible
Bug 32864127

When a fragment creates a view that is either invisible or gone,
it should remain that way. This means that the fragment won't have
any animations. When transactions are optimized, transitions
will avoid invisible and gone fragment views, but unoptimized
transactions will lead to failed transitions because none of the
views will be visible.

Test: Ie71902d1b5f5f647c758a1bd60cdfd6814cda29d
Change-Id: If952f81859227a28569d234d8232ff6bb100e74c

Former-commit-id: 0c3ab75c6a9da73043cd345d05a3d1e47c455e69