Hold a wake lock while DeviceIdleController is going idle.
The inputs to DeviceIdleController (alarm manager, sensors)
hold wake locks while they call it.  But then the real work
happens in a handler which is outside of the wakelock, so
listeners don't get a chance to run right away, which in
the case of NetworkPolicyManager means the device is in a
higher power state than it should be.

It's not clear that this will 100% fix the bug, because
NetworkPolicyManagerService also has its own internal
Handler, and isn't holding its own wakelock for this,
but this change allows NPMS to be fixed if it really
needed to be.

Bug: 31900521
Test: adb shell dumpsys deviceidle step ... adb shell dumpsys power has no wakelocks held
Change-Id: I799f45221a6b327e7f63745cadc95ca644c064c6

Former-commit-id: 8f0e9cede145bbc1668e25cfd6f850fe3fde3ae3