Add support for audio-focused notifications.
Add extras keys and methos to Notification and RemoteInput for notifications
whose contents have an audio representation and whose RemoteInputs
can accept data results (useful for, e.g., voice messaging
applications, photo results, etc).

This motivator for this change is it enables an end-to-end flow for
audio messaging including
  * System receives the audio message
  * Replying to the message with the user's voice
  * App receives the audio reply data

Note that this CL just adds the capabilities, but does not
add any UI or implementation to perform the above flow.

Test: As agreed, this will be tested both with a set of unit tests,
      and with an "end-to-end" test that simulates two apps, one
      creating an audio message with audio reply action, and a
      second app that executes that action, providing an audio
      result.

Change-Id: I55779ac4b781d8273c13c69fded0b56ee639cf01

Former-commit-id: 472a3b330a926c9d86a685dc0debf71037ea279b